// frontend/src/app/models/User.ts
export class User{
    username: string="";
    password: string="";
    firstname: string="";
    lastname: string="";
    gender: string="";
    address: string="";
    phone: string="";
    email: string="";
    picture: string="";
    card: string="";
    type: string="";
    state: string="waiting";
    haveFirm: boolean=false;
    isWorking: boolean=false;
    firmName: string="";
}